requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.4.2
google-api-python-client==2.128.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
streamlit==1.35.0
pandas==2.2.2
openai==1.28.1
selenium==4.21.0
webdriver-manager==4.0.1
python-dateutil==2.8.2
